requirements/req-main.txt
# SHA1:a21c9c7bf6621600c842576a2abd8bcbe7c68ae9
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
lxml==4.6.3
    # via pikepdf
pikepdf==2.12.0
    # via -r requirements/req-main.in
pillow==8.2.0
    # via pikepdf
pycairo==1.20.0
    # via -r requirements/req-main.in
